create,delete,update post

+++ routes/users.js
@@ -19,7 +19,7 @@ router.put("/:id", async (req, res) => {
     const user = await User.findByIdAndUpdate(req.params.id, {
       $set: req.body,
     });
-    res.status(200).json({ message: "Account has been updated!" });
+    return res.status(200).json({ message: "Account has been updated!" });
   } else {
     return res
       .send(403)
@@ -34,9 +34,11 @@ router.delete("/:id", async (req, res) => {
   ) {
     try {
       const user = await User.findOne({ _id: req.params.id });
-      !user &&
-        res.status(404).json({ message: "Sorry! the User is not available" });
-
+      if (!user) {
+        return res
+          .status(404)
+          .json({ message: "Sorry! the User is not available" });
+      }
       await User.deleteOne({ _id: req.params.id }); //or
       // await User.findByIdAndDelete(req.params.id);
       res.status(200).json({ message: "Account has been deleted!" });
@@ -82,15 +84,17 @@ router.put("/:id/follow", async (req, res) => {
         await user.updateOne({ $push: { followers: req.body.user_id } });
         await authUser.updateOne({ $push: { followings: req.params.id } });
 
-        res.status(200).json({ message: `You followed ${authUser.username}` });
+        return res
+          .status(200)
+          .json({ message: `You followed ${authUser.username}` });
       } else {
-        res.status(403).json({ message: "You already follow the user" });
+        return res.status(403).json({ message: "You already follow the user" });
       }
     } catch (err) {
-      res.status(500).json(err.message);
+      return res.status(500).json(err.message);
     }
   } else {
-    res.status(403).json("You can't follow yourself");
+    return res.status(403).json("You can't follow yourself");
   }
 });
 // Unfollow a user
@@ -102,17 +106,19 @@ router.put("/:id/unfollow", async (req, res) => {
       if (user.followers.includes(req.body.user_id)) {
         await user.updateOne({ $pull: { followers: req.body.user_id } });
         await authUser.updateOne({ $pull: { followings: req.params.id } });
-        res
+        return res
           .status(200)
           .json({ message: `You just unfollow ${authUser.username}` });
       } else {
-        res.status(403).json({ message: "You don't follow the user before" });
+        return res
+          .status(403)
+          .json({ message: "You don't follow the user before" });
       }
     } catch (err) {
-      res.status(500).json(err.message);
+      return res.status(500).json(err.message);
     }
   } else {
-    res.status(403).json("You can't unfollow yourself");
+    return res.status(403).json("You can't unfollow yourself");
   }
 });
